fix(topic_entity_tag): fixed reference-related creation and tests + moved source to a separate test
